move class DocValidator to package pytools.api and add docstrings
