+++ src/pytools/api/__init__.py
@@ -3,3 +3,4 @@ Basic tools for API development, supporting documentation, deprecation,
 and run-time validation.
 """
 from ._api import *
+from ._doc_validator import *
